Move types to a separate file within each package
This makes it easier to read. Most developers have LSP set up so this should be fine
