+++ cmd/gandi.go
@@ -22,8 +22,8 @@ type cli struct {
 }
 
 type globals struct {
-	liveDNSHandle *livedns.LiveDNS
-	domainHandle  *domain.Domain
+	liveDNSHandle *livedns.LiveDNSAPI
+	domainHandle  *domain.DomainAPI
 	Version       versionFlag `kong:"name='version',help='Print version information and quit'"`
 }
