+++ livedns/livedns.go
@@ -4,20 +4,15 @@ import (
 	"github.com/go-gandi/go-gandi/internal/client"
 )
 
-// LiveDNS is the API client to the Gandi v5 LiveDNS API
-type LiveDNS struct {
-	client client.Gandi
-}
-
 // New returns an instance of the LiveDNS API client
-func New(apikey string, sharingid string, debug bool, dryRun bool) *LiveDNS {
+func New(apikey string, sharingid string, debug bool, dryRun bool) *LiveDNSAPI {
 	client := client.New(apikey, sharingid, debug, dryRun)
 	client.SetEndpoint("livedns/")
-	return &LiveDNS{client: *client}
+	return &LiveDNSAPI{client: *client}
 }
 
 // NewFromClient returns an instance of the LiveDNS API client
-func NewFromClient(g client.Gandi) *LiveDNS {
+func NewFromClient(g client.Gandi) *LiveDNSAPI {
 	g.SetEndpoint("livedns/")
-	return &LiveDNS{client: g}
+	return &LiveDNSAPI{client: g}
 }
